requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.2
deprecation==2.1.0
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dj-database-url==0.5.0
frozenlist==1.6.0
gotrue==2.11.1
gunicorn==20.1.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
multidict==6.4.3
packaging==25.0
paypalrestsdk==1.13.3
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
psycopg2==2.9.10
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.8.0
pyOpenSSL==25.1.0
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
realtime==2.4.3
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.1
supafunc==0.9.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websockets==14.2
whitenoise==6.2.0
yarl==1.20.0